+++ src/main/java/com/pmu2/exec/infrastrure/db/sql/PartantJpaRepository.java
@@ -1,11 +1,10 @@
-package com.pmu2.exec.infrastrure.inbound;
+package com.pmu2.exec.infrastrure.db.sql;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface PartantJpaRepository extends JpaRepository<PartantEntity, Long> {
